requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
faiss-cpu>=1.7.4
tavily-python>=0.3.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
plotly>=5.17.0
streamlit-aggrid>=0.3.4
requests>=2.31.0

